character movement and puzzle level
I changed the player movement script slightly to allow the player to fall through and jump through platforms. If you want this functionality, put the "Platform" tag on a platform, along with a platform effector 2d component. Check the box on the platforms box collider 2d component that says used by effector then it should work