Dialogue Box Pop-ups
I have added 3 new scripts to implement dialogue boxes popping up in the current tutorial LGTL-J as well as an animation of it sliding up rather than popping up instantaneously. Next update, I want to add in the dialogue scripts and triggers, possibly a character for the professor in the dialogue window, and add animations for the truth tables and logic gates as the player goes through the tutorial.