Tutorial Level Finished
Tutorial level has been completed. Things to note: I have not added the text highlight yet, will add in next update. The platform movement script is also bugging a little bit, it won't let you walk normally when on the platform so I'll look into that. Also the dialogue box looks different than what was shown in class because the animator is not showing the full thing for some reason, i will try to fix that in the next update as well.